requirements.txt
# ============================================================================
# DEPENDÊNCIAS DO BOT MEV
# ============================================================================

# Blockchain e Web3
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# HTTP e APIs
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Machine Learning e IA
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Dashboard Web
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5

# Utilitários
python-dotenv==1.0.0
python-telegram-bot==20.7
colorama==0.4.6
tqdm==4.66.1

# Logging e Monitoramento
loguru==0.7.2

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
